+++ src/Aspire.Hosting.LocalStack/Internal/Constants.cs
@@ -6,4 +6,7 @@ internal static class Constants
     internal const int DefaultContainerPort = 4566;
     internal const string CloudFormationReferenceAnnotation = "Aspire.Hosting.AWS.CloudFormation.CloudFormationReferenceAnnotation";
     internal const string SQSEventSourceResource = "Aspire.Hosting.AWS.Lambda.SQSEventSourceResource";
+
+    internal const string LocalStackHealthClientName = "localstack_health_client";
+    internal const string LocalStackHealthCheckName = "localstack_health";
 }